Update manifest Ifdef x86 Eigen CSR sparse matmul implementation Exclude broken test, check for string type when copying cross device Add __wasm__ conditional Split pydind schema, regenerate docs, add exclusion Conditionally exclude schema module

+++ onnxruntime/core/framework/sparse_utils.cc
@@ -75,7 +75,7 @@ common::Status DenseTensorToSparseCsr(const DataTransferManager& data_manager, c
   std::vector<int64_t> inner_indicies;
   inner_indicies.reserve(dense_elements / 2);
   std::vector<int64_t> outer_indices;
-  outer_indices.reserve(rows + 1);
+  outer_indices.reserve(static_cast<size_t>(rows) + 1);
 
   Tensor nnz_tensor;
   std::vector<uint8_t> values_8;
